+++ azurefunctions-extensions-bindings-cosmosdb/LICENSE
