+++ azurefunctions-extensions-bindings-cosmosdb/MANIFEST.in
