+++ azurefunctions-extensions-bindings-cosmosdb/azurefunctions/__init__.py
