+++ azurefunctions-extensions-bindings-cosmosdb/azurefunctions/extensions/__init__.py
