+++ azurefunctions-extensions-bindings-cosmosdb/azurefunctions/extensions/bindings/__init__.py
