+++ azurefunctions-extensions-bindings-cosmosdb/azurefunctions/extensions/bindings/cosmosdb/containerProxy.py
