+++ azurefunctions-extensions-bindings-cosmosdb/azurefunctions/extensions/bindings/cosmosdb/cosmosClient.py
