+++ azurefunctions-extensions-bindings-cosmosdb/azurefunctions/extensions/bindings/cosmosdb/cosmosClientConverter.py
