+++ azurefunctions-extensions-bindings-cosmosdb/azurefunctions/extensions/bindings/cosmosdb/databaseProxy.py
