+++ azurefunctions-extensions-bindings-cosmosdb/azurefunctions/extensions/bindings/cosmosdb/utils.py
