+++ azurefunctions-extensions-bindings-cosmosdb/samples/cosmosdb_samples_containerproxy/host.json
