+++ azurefunctions-extensions-bindings-cosmosdb/samples/cosmosdb_samples_cosmosclient/host.json
