+++ azurefunctions-extensions-bindings-cosmosdb/samples/cosmosdb_samples_cosmosclient/local.settings.json
@@ -2,6 +2,6 @@
   "IsEncrypted": false,
   "Values": {
     "FUNCTIONS_WORKER_RUNTIME": "python",
-    "AzureWebJobsStorage": "UseDevelopmentStorage=true"
+    "CosmosDBConnection": ""
   }
 }
