+++ azurefunctions-extensions-bindings-cosmosdb/samples/cosmosdb_samples_cosmosclient/requirements.txt
@@ -3,4 +3,4 @@
 # Manually managing azure-functions-worker may cause unexpected issues
 
 azure-functions
-azurefunctions-extensions-bindings-cosmos
+azurefunctions-extensions-bindings-cosmosdb
