+++ azurefunctions-extensions-bindings-cosmosdb/samples/cosmosdb_samples_databaseproxy/host.json
