+++ eng/ci/official-build.yml
@@ -51,10 +51,10 @@ extends:
         dependsOn: Build
         jobs:
           - template: /eng/templates/official/jobs/blob-unit-tests.yml@self
-      - stage: RunCosmosTests
+      - stage: RunCosmosDBTests
         dependsOn: Build
         jobs:
-          - template: /eng/templates/official/jobs/cosmos-unit-tests.yml@self
+          - template: /eng/templates/official/jobs/cosmosdb-unit-tests.yml@self
       - stage: RunFastApiTests
         dependsOn: Build
         jobs:
